LPS-63898 Copy getDDMXSD operation to UpgradeJournal
