+++ modules/apps/web-experience/journal/journal-api/src/main/java/com/liferay/journal/util/JournalConverter.java
@@ -20,8 +20,6 @@ import com.liferay.dynamic.data.mapping.storage.Fields;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.xml.Document;
 
-import java.util.Locale;
-
 /**
  * @author Marcellus Tavares
  * @author Bruno Basto
@@ -38,14 +36,6 @@ public interface JournalConverter {
 		throws PortalException;
 
 	public DDMFormValues getDDMFormValues(
-			DDMStructure ddmStructure, Fields fields)
-		throws PortalException;
-
-	public String getDDMXSD(String journalXSD) throws Exception;
-
-	public String getDDMXSD(String journalXSD, Locale defaultLocale)
-		throws Exception;
-
-	public String getJournalXSD(String ddmXSD) throws Exception;
+		DDMStructure ddmStructure, Fields fields)throws PortalException;
 
 }
