+++ modules/apps/web-experience/journal/journal-test/src/testIntegration/java/com/liferay/journal/util/test/JournalConverterUtilTest.java
@@ -33,7 +33,6 @@ import com.liferay.dynamic.data.mapping.util.DDMXML;
 import com.liferay.dynamic.data.mapping.util.impl.DDMImpl;
 import com.liferay.dynamic.data.mapping.util.impl.DDMXMLImpl;
 import com.liferay.journal.model.JournalArticle;
-import com.liferay.journal.test.util.JournalTestUtil;
 import com.liferay.journal.util.JournalConverter;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
@@ -315,23 +314,6 @@ public class JournalConverterUtilTest {
 	}
 
 	@Test
-	public void testGetDDMXSD() throws Exception {
-		String expectedXSD = read("test-ddm-structure-all-fields.xml");
-
-		DDMForm expectedDDMForm = _ddmFormXSDDeserializer.deserialize(
-			expectedXSD);
-
-		String actualXSD = _journalConverter.getDDMXSD(
-			read("test-journal-structure-all-fields.xml"));
-
-		validateDDMXSD(actualXSD);
-
-		DDMForm actualDDMForm = _ddmFormXSDDeserializer.deserialize(actualXSD);
-
-		assertEquals(expectedDDMForm, actualDDMForm);
-	}
-
-	@Test
 	public void testGetFieldsFromContentWithBooleanElement() throws Exception {
 		Fields expectedFields = new Fields();
 
@@ -528,22 +510,6 @@ public class JournalConverterUtilTest {
 			_ddmStructure, content);
 
 		Assert.assertEquals(expectedFields, actualFields);
-	}
-
-	@Test
-	public void testGetJournalXSD() throws Exception {
-		String expectedXSD = read("test-journal-structure-all-fields.xml");
-
-		Map<String, Map<String, String>> expectedMap =
-			JournalTestUtil.getXsdMap(expectedXSD);
-
-		String actualXSD = _journalConverter.getJournalXSD(
-			read("test-ddm-structure-all-fields.xml"));
-
-		Map<String, Map<String, String>> actualMap = JournalTestUtil.getXsdMap(
-			actualXSD);
-
-		Assert.assertEquals(expectedMap, actualMap);
 	}
 
 	protected void assertEquals(
