+++ portal-impl/src/com/liferay/portal/upgrade/v6_2_0/UpgradeJournal.java
@@ -14,12 +14,13 @@
 
 package com.liferay.portal.upgrade.v6_2_0;
 
-import com.liferay.journal.kernel.util.JournalConverterManagerUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.portlet.PortletPreferencesFactoryUtil;
 import com.liferay.portal.kernel.upgrade.BaseUpgradePortletPreferences;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.HtmlUtil;
+import com.liferay.portal.kernel.util.HttpUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.LoggingTimer;
 import com.liferay.portal.kernel.util.PortalUtil;
@@ -27,14 +28,21 @@ import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.xml.Attribute;
+import com.liferay.portal.kernel.xml.Document;
+import com.liferay.portal.kernel.xml.Element;
+import com.liferay.portal.kernel.xml.SAXReaderUtil;
+import com.liferay.portal.kernel.xml.XPath;
 import com.liferay.portal.upgrade.AutoBatchPreparedStatementUtil;
 import com.liferay.portal.upgrade.v6_2_0.util.JournalFeedTable;
+import com.liferay.util.xml.XMLUtil;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Timestamp;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
@@ -80,10 +88,7 @@ public class UpgradeJournal extends BaseUpgradePortletPreferences {
 			ps.setString(11, ddmStructureKey);
 			ps.setString(12, name);
 			ps.setString(13, description);
-			ps.setString(
-				14,
-				JournalConverterManagerUtil.getDDMXSD(
-					xsd, getDefaultLocale(companyId)));
+			ps.setString(14, getDDMXSD(xsd, getDefaultLocale(companyId)));
 			ps.setString(15, storageType);
 			ps.setInt(16, type);
 
@@ -179,6 +184,24 @@ public class UpgradeJournal extends BaseUpgradePortletPreferences {
 		}
 	}
 
+	protected void addMetadataEntry(
+		Element metadataElement, String name, String value) {
+
+		Element entryElement = metadataElement.addElement("entry");
+
+		entryElement.addAttribute("name", name);
+		entryElement.addCDATA(value);
+	}
+
+	protected String decodeURL(String url) {
+		try {
+			return HttpUtil.decodeURL(url);
+		}
+		catch (IllegalArgumentException iae) {
+			return url;
+		}
+	}
+
 	@Override
 	protected void doUpgrade() throws Exception {
 		alter(
@@ -191,6 +214,22 @@ public class UpgradeJournal extends BaseUpgradePortletPreferences {
 		updateAssetEntryClassTypeId();
 
 		super.doUpgrade();
+	}
+
+	protected Element fetchMetadataEntry(
+		Element parentElement, String attributeName, String attributeValue) {
+
+		StringBundler sb = new StringBundler(5);
+
+		sb.append("entry[@");
+		sb.append(attributeName);
+		sb.append(StringPool.EQUAL);
+		sb.append(HtmlUtil.escapeXPathAttribute(attributeValue));
+		sb.append(StringPool.CLOSE_BRACKET);
+
+		XPath xPathSelector = SAXReaderUtil.createXPath(sb.toString());
+
+		return (Element)xPathSelector.selectSingleNode(parentElement);
 	}
 
 	protected long getCompanyGroupId(long companyId) throws Exception {
@@ -267,6 +306,27 @@ public class UpgradeJournal extends BaseUpgradePortletPreferences {
 		return getDDMStructureId(groupId, 0, structureId, warn);
 	}
 
+	protected String getDDMXSD(String journalXSD, Locale defaultLocale)
+		throws Exception {
+
+		Document document = SAXReaderUtil.read(journalXSD);
+
+		Element rootElement = document.getRootElement();
+
+		rootElement.addAttribute("available-locales", defaultLocale.toString());
+		rootElement.addAttribute("default-locale", defaultLocale.toString());
+
+		List<Element> dynamicElementElements = rootElement.elements(
+			"dynamic-element");
+
+		for (Element dynamicElementElement : dynamicElementElements) {
+			updateJournalXSDDynamicElement(
+				dynamicElementElement, defaultLocale.toString());
+		}
+
+		return XMLUtil.formatXML(document);
+	}
+
 	protected Locale getDefaultLocale(long companyId) throws Exception {
 		try (PreparedStatement ps = connection.prepareStatement(
 				"select languageId from User_ where companyId = ? and " +
@@ -292,6 +352,16 @@ public class UpgradeJournal extends BaseUpgradePortletPreferences {
 		return new String[] {
 			"56_INSTANCE_%", "62_INSTANCE_%", "101_INSTANCE_%"
 		};
+	}
+
+	protected void removeAttribute(Element element, String attributeName) {
+		Attribute attribute = element.attribute(attributeName);
+
+		if (attribute == null) {
+			return;
+		}
+
+		element.remove(attribute);
 	}
 
 	protected void updateAssetEntryClassTypeId() throws Exception {
@@ -328,6 +398,168 @@ public class UpgradeJournal extends BaseUpgradePortletPreferences {
 
 				ps2.executeBatch();
 			}
+		}
+	}
+
+	protected void updateJournalXSDDynamicElement(
+		Element element, String defaultLanguageId) {
+
+		Map<String, String> _ddmDataTypes = new HashMap<>();
+
+		_ddmDataTypes.put("boolean", "boolean");
+		_ddmDataTypes.put("document_library", "document-library");
+		_ddmDataTypes.put("image", "image");
+		_ddmDataTypes.put("link_to_layout", "link-to-page");
+		_ddmDataTypes.put("list", "string");
+		_ddmDataTypes.put("multi-list", "string");
+		_ddmDataTypes.put("text", "string");
+		_ddmDataTypes.put("text_area", "html");
+		_ddmDataTypes.put("text_box", "string");
+
+		Map<String, String> _ddmMetadataAttributes = new HashMap<>();
+
+		_ddmMetadataAttributes.put("instructions", "tip");
+		_ddmMetadataAttributes.put("label", "label");
+		_ddmMetadataAttributes.put("predefinedValue", "predefinedValue");
+
+		Map<String, String> _journalTypesToDDMTypes = new HashMap<>();
+
+		_journalTypesToDDMTypes.put("boolean", "checkbox");
+		_journalTypesToDDMTypes.put("document_library", "ddm-documentlibrary");
+		_journalTypesToDDMTypes.put("image", "ddm-image");
+		_journalTypesToDDMTypes.put("image_gallery", "ddm-documentlibrary");
+		_journalTypesToDDMTypes.put("link_to_layout", "ddm-link-to-page");
+		_journalTypesToDDMTypes.put("list", "select");
+		_journalTypesToDDMTypes.put("multi-list", "select");
+		_journalTypesToDDMTypes.put("selection_break", "ddm-separator");
+		_journalTypesToDDMTypes.put("text", "text");
+		_journalTypesToDDMTypes.put("text_area", "ddm-text-html");
+		_journalTypesToDDMTypes.put("text_box", "textarea");
+
+		String name = element.attributeValue("name");
+		String type = element.attributeValue("type");
+
+		Element metadataElement = element.element("meta-data");
+
+		if (metadataElement == null) {
+			metadataElement = element.addElement("meta-data");
+		}
+
+		if (type.equals("multi-list")) {
+			element.addAttribute("multiple", "true");
+		}
+		else {
+			Element parentElement = element.getParent();
+
+			String parentType = parentElement.attributeValue("type");
+
+			if ((parentType != null) && parentType.equals("select")) {
+				metadataElement.addAttribute("locale", defaultLanguageId);
+
+				addMetadataEntry(metadataElement, "label", decodeURL(name));
+
+				removeAttribute(element, "index-type");
+
+				element.addAttribute("name", "option" + StringUtil.randomId());
+				element.addAttribute("type", "option");
+				element.addAttribute("value", decodeURL(type));
+
+				return;
+			}
+		}
+
+		String indexType = StringPool.BLANK;
+
+		Attribute indexTypeAttribute = element.attribute("index-type");
+
+		if (indexTypeAttribute != null) {
+			indexType = indexTypeAttribute.getValue();
+
+			element.remove(indexTypeAttribute);
+		}
+
+		element.remove(element.attribute("type"));
+
+		if (!type.equals("selection_break")) {
+			String dataType = _ddmDataTypes.get(type);
+
+			if (dataType == null) {
+				dataType = "string";
+			}
+
+			element.addAttribute("dataType", dataType);
+		}
+
+		element.addAttribute("indexType", indexType);
+
+		String required = "false";
+
+		Element requiredElement = fetchMetadataEntry(
+			metadataElement, "name", "required");
+
+		if (requiredElement != null) {
+			required = requiredElement.getText();
+		}
+
+		element.addAttribute("required", required);
+
+		element.addAttribute("showLabel", "true");
+
+		String newType = _journalTypesToDDMTypes.get(type);
+
+		if (newType == null) {
+			newType = type;
+		}
+
+		element.addAttribute("type", newType);
+
+		if (newType.startsWith("ddm")) {
+			element.addAttribute("fieldNamespace", "ddm");
+		}
+
+		metadataElement.addAttribute("locale", defaultLanguageId);
+
+		List<Element> entryElements = metadataElement.elements();
+
+		if (entryElements.isEmpty()) {
+			addMetadataEntry(metadataElement, "label", name);
+		}
+		else {
+			for (Element entryElement : entryElements) {
+				String oldEntryName = entryElement.attributeValue("name");
+
+				String newEntryName = _ddmMetadataAttributes.get(oldEntryName);
+
+				if (newEntryName == null) {
+					metadataElement.remove(entryElement);
+				}
+				else {
+					entryElement.addAttribute("name", newEntryName);
+				}
+			}
+		}
+
+		if (newType.equals("ddm-date") || newType.equals("ddm-decimal") ||
+			newType.equals("ddm-integer") ||
+			newType.equals("ddm-link-to-page") ||
+			newType.equals("ddm-number") || newType.equals("ddm-text-html") ||
+			newType.equals("text") || newType.equals("textarea")) {
+
+			element.addAttribute("width", "25");
+		}
+		else if (newType.equals("ddm-image")) {
+			element.addAttribute("fieldNamespace", "ddm");
+			element.addAttribute("readOnly", "false");
+		}
+
+		element.add(metadataElement.detach());
+
+		List<Element> dynamicElementElements = element.elements(
+			"dynamic-element");
+
+		for (Element dynamicElementElement : dynamicElementElements) {
+			updateJournalXSDDynamicElement(
+				dynamicElementElement, defaultLanguageId);
 		}
 	}
